backend/requirements.txt
# Core FastAPI and server dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.0
pydantic>=2.7.2
pydantic-settings>=2.0.0
httpx>=0.26.0,<0.29.0

# Database and authentication
supabase>=2.15.2
PyJWT>=2.10.1,<3.0.0

# Google AI and authentication dependencies
google-generativeai>=0.8.3
google-cloud-aiplatform>=1.87.0
google-auth>=2.29.0
gotrue>=2.12.0

# OpenAI integration
openai>=1.0.0

# Anthropic integration
anthropic>=0.40.0

# Validation and utilities
email-validator>=2.1.0

# Testing dependencies
pytest>=7.4.3
pytest-asyncio>=0.21.1

# HTTP and API dependencies
aiohttp>=3.9.1
beautifulsoup4>=4.12.2 